Ajout de la classe LinksPanel permettant d'afficher les liens
- Lorsque l'on appuie sur le bouton "Afficher les liens hypertextes" les autres composans en dessous du titre se cachent pour laisser place aux liens hypertextes

+++ src/fr/r34/metagg/gui/MainMenuGUI.java
@@ -87,6 +87,7 @@ public class MainMenuGUI {
                 );
                 if (response == JOptionPane.OK_OPTION) {
                     for (MetaFile metaFile : metaFilesOpened) {
+                        if (!metaFile.getDestDir().exists()) continue;
                         metaFile.save();
                         metaFile.deleteTempFolder();
                     }
